feat: 큰 단위 컴포넌트 분리작업
