+++ src/components/OverlapDialog.tsx
@@ -1,0 +1,48 @@
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogContentText,
+  DialogTitle,
+  Typography,
+} from '@mui/material';
+
+import type { Event } from '../types';
+
+type OverlapDialogProps = {
+  isOpen: boolean;
+  overlappingEvents: Event[];
+  onConfirm: () => void;
+  onCancel: () => void;
+};
+
+export function OverlapDialog({
+  isOpen,
+  overlappingEvents,
+  onConfirm,
+  onCancel,
+}: OverlapDialogProps) {
+  return (
+    <Dialog open={isOpen} onClose={onCancel}>
+      <DialogTitle>일정 겹침 경고</DialogTitle>
+      <DialogContent>
+        <DialogContentText>
+          다음 일정과 겹칩니다:
+          {overlappingEvents.map((event) => (
+            <Typography key={event.id}>
+              {event.title} ({event.date} {event.startTime}-{event.endTime})
+            </Typography>
+          ))}
+          계속 진행하시겠습니까?
+        </DialogContentText>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onCancel}>취소</Button>
+        <Button color="error" onClick={onConfirm}>
+          계속 진행
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
