+++ src/main.tsx
@@ -4,7 +4,7 @@ import { SnackbarProvider } from 'notistack';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App.tsx';
+import { App } from './App';
 
 const theme = createTheme();
